Errorhandler refactored

+++ cms.pp/src/main/java/com/cms/pp/cms/pp/Priviliges/PrivilegeRepository.java
@@ -2,6 +2,6 @@ package com.cms.pp.cms.pp.Priviliges;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PrivilegeRepository extends JpaRepository<Privilege, Integer> {
+public interface PrivilegeRepository extends JpaRepository<Privilege, Long> {
     Privilege findByName(String name);
 }
